requirements.txt
discord.py
python-dotenv
email-validator
aiosmtplib
aiohttp
